+++ packages/create-agents/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @inkeep/create-agents
 
+## 0.69.0
+
+### Patch Changes
+
+- Updated dependencies [52d0831]
+- Updated dependencies [c63567e]
+- Updated dependencies [32bce4f]
+  - @inkeep/agents-core@0.69.0
+
 ## 0.68.4
 
 ### Patch Changes
@@ -1508,7 +1517,6 @@
 
 - dba5a31: Update quickstart port check
 - b0817aa: Fix CLI bugs
-
   - Quickstart inkeep.config.ts indents and types
   - inkeep init run API and manage API urls